+++ src/Pinboard/Client.hs
@@ -152,7 +152,7 @@ buildReq :: MonadIO m => String -> m Request
 buildReq url = do
   req <- liftIO $ parseUrl $ "https://api.pinboard.in/v1/" <> url
   return $ req 
-    { requestHeaders = [("User-Agent","pinboard.hs/0.9.5")]
+    { requestHeaders = [("User-Agent","pinboard.hs/0.9.6")]
     , checkStatus = \_ _ _ -> Nothing
     }