+++ changelog.md
@@ -1,6 +1,6 @@
 __v0.9.5__
 
-
+add JSON roundtrip tests
 
 __v0.9.4__