+++ changelog.md
@@ -1,3 +1,7 @@
+__v0.9.6__
+
+add http-client bound
+
 __v0.9.5__
 
 add JSON roundtrip tests